Azrael Weapon Assets
The weapons for our player character.
NOTE: I have made a separate material in the GLTF format, located under the WeaponTextures folder. When importing weapons, if you want to use my material, be sure that the option to "Create New Material" is set to "Do Not Make New Material", and import the GLTF separately.